+++ api/src/test/java/it/unicam/cs/pa/ma114763/logo/statement/SetStrokeColorStatementTest.java
@@ -24,8 +24,8 @@ class SetStrokeColorStatementTest {
         assertEquals(2, statements.size());
         assertInstanceOf(SetStrokeColorStatement.class, statements.get(0));
         assertInstanceOf(SetStrokeColorStatement.class, statements.get(1));
-        TestUtils.assertColorStatement((SetStrokeColorStatement) statements.get(0), (byte) 0, (byte) 0, (byte) 0, (byte) 255);
-        TestUtils.assertColorStatement((SetStrokeColorStatement) statements.get(1), (byte) 10, (byte) 20, (byte) 30, (byte) 67);
+        TestUtils.assertColorStatement((SetStrokeColorStatement) statements.get(0), 0, 0, 0, 255);
+        TestUtils.assertColorStatement((SetStrokeColorStatement) statements.get(1), 10, 20, 30, 67);
     }
 
     @Test
